1. Added MapboxView and InputView to view folder. 2. Refactored code files to use MVC architecture. 3. Updated package.json file.
