+++ mapbox/package.json
@@ -4,6 +4,7 @@
     "private": true,
     "dependencies": {
       "mapbox-gl": "^2.9.2",
+      "react-map-gl-geocoder": "^2.2.0",
       "react": "^17.0.2",
       "react-dom": "^17.0.2",
       "react-scripts": "^4.0.3"
