+++ mapbox/src/App.css
@@ -1,0 +1,3 @@
+.main {
+    display: flex;
+  }
