+++ mapbox/src/view/MapboxView.css
@@ -1,0 +1,26 @@
+.map-container {
+    width: 70vw;
+    height: 100vh;
+    /* position: absolute;
+    top: 0;
+    bottom: 0;
+    width: 100%; */
+  }
+  
+
+.sidebar {
+background-color: rgba(35, 55, 75, 0.9);
+color: #fff;
+padding: 6px 12px;
+font-family: monospace;
+z-index: 1;
+position: absolute;
+top: 0;
+left: 0;
+margin: 12px;
+border-radius: 4px;
+}
+
+.mapView {
+  display: flex;
+}
